+++ planetiler-core/src/test/java/com/onthegomap/planetiler/collection/FeatureGroupTest.java
@@ -91,7 +91,7 @@ class FeatureGroupTest {
     Map<Integer, Map<String, List<Feature>>> map = new TreeMap<>();
     for (FeatureGroup.TileFeatures tile : features) {
       for (var feature : VectorTile.decode(tile.getVectorTileEncoder().encode())) {
-        map.computeIfAbsent(tile.tileCoord().encoded(), (i) -> new TreeMap<>())
+        map.computeIfAbsent((int)tile.tileCoord().encoded(), (i) -> new TreeMap<>())
           .computeIfAbsent(feature.layer(), l -> new ArrayList<>())
           .add(new Feature(feature.attrs(), decodeSilently(feature.geometry())));
       }
@@ -105,7 +105,7 @@ class FeatureGroupTest {
     var reader = features.parallelIterator(2);
     for (FeatureGroup.TileFeatures tile : reader.result()) {
       for (var feature : VectorTile.decode(tile.getVectorTileEncoder().encode())) {
-        map.computeIfAbsent(tile.tileCoord().encoded(), (i) -> new TreeMap<>())
+        map.computeIfAbsent((int)tile.tileCoord().encoded(), (i) -> new TreeMap<>())
           .computeIfAbsent(feature.layer(), l -> new ArrayList<>())
           .add(new Feature(feature.attrs(), decodeSilently(feature.geometry())));
       }
@@ -391,7 +391,7 @@ class FeatureGroupTest {
       for (byte layer : layers) {
         for (int sortKey : sortKeys) {
           for (boolean hasGroup : hasGroups) {
-            long key = FeatureGroup.encodeKey(tile.encoded(), layer, sortKey, hasGroup);
+            long key = FeatureGroup.encodeKey((int)tile.encoded(), layer, sortKey, hasGroup);
             result.add(dynamicTest(tile + " " + layer + " " + sortKey + " " + hasGroup, () -> {
               assertEquals(tile.encoded(), FeatureGroup.extractTileFromKey(key), "tile");
               assertEquals(layer, FeatureGroup.extractLayerIdFromKey(key), "layer");
